+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Feat: Add tracestate HTTP header support (#1683)
 * Feat: Add option to filter which origins receive tracing headers (#1698)
+* Feat: Include unfinished spans in transaction (#1699)
 
 Breaking changes: